requirements.txt
llama-index-llms-openai
python-dotenv
langchain-openai
langchain_community
duckduckgo-search